+++ lib/config.js
@@ -2,7 +2,8 @@
 
 import BLOG from '@/blog.config'
 import { useGlobal } from './global'
-import { deepClone, isUrl } from './utils'
+import { deepClone } from './utils'
+import { isUrl } from '@/lib/utils/url'
 
 /**
  * 读取配置顺序
