+++ lib/lang.js
@@ -1,5 +1,5 @@
 import BLOG from '@/blog.config'
-import { getQueryVariable, isBrowser, mergeDeep } from '@/lib/utils'
+import { isBrowser, mergeDeep } from '@/lib/utils'
 import enUS from './lang/en-US'
 import frFR from './lang/fr-FR'
 import jaJP from './lang/ja-JP'
@@ -8,7 +8,8 @@ import zhCN from './lang/zh-CN'
 import zhHK from './lang/zh-HK'
 import zhTW from './lang/zh-TW'
 import { extractLangPrefix } from './utils/pageId'
-import { useRouter } from 'next/router'
+
+import { getQueryVariable } from '@/lib/utils/url'
 
 /**
  * 在这里配置所有支持的语言
