+++ lib/notion/convertInnerUrl.js
@@ -1,5 +1,6 @@
 import { idToUuid } from 'notion-utils'
-import { checkStrIsNotionId, getLastPartOfUrl, isBrowser } from '../utils'
+import { checkStrIsNotionId, isBrowser } from '../utils'
+import { getLastPartOfUrl } from '@/lib/utils/url'
 
 /**
  * 处理页面内连接跳转:
