+++ lib/notion/getPageProperties.js
@@ -4,14 +4,11 @@ import formatDate from '../utils/formatDate'
 // import { createHash } from 'crypto'
 import md5 from 'js-md5'
 import { siteConfig } from '../config'
-import {
-  checkStartWithHttp,
-  convertUrlStartWithOneSlash,
-  getLastSegmentFromUrl
-} from '../utils'
+import { checkStartWithHttp } from '../utils'
 import { extractLangPrefix } from '../utils/pageId'
 import { mapImgUrl } from './mapImage'
 import notionAPI from '@/lib/notion/getNotionAPI'
+import { convertUrlStartWithOneSlash, getLastSegmentFromUrl } from '@/lib/utils/url'
 
 /**
  * 获取页面元素成员属性
